requirements.txt
google-genai>=0.3.0
psycopg2-binary>=2.9.9
tqdm>=4.66.2
python-dotenv>=1.0.1
